Rough migration of distributed tensor out
