+++ scripts/validate_api/validate_includes.py
@@ -52,8 +52,8 @@ EXPERIMENTAL_TENSOR_FORWARD_HEADERS = {
     "tt-metalium/experimental/tensor/spec/layout/page_config.hpp": "ttnn/tensor/layout/page_config.hpp",
     "tt-metalium/experimental/tensor/spec/layout/tensor_layout.hpp": "ttnn/tensor/layout/tensor_layout.hpp",
     "tt-metalium/experimental/tensor/spec/memory_config/memory_config.hpp": "ttnn/tensor/memory_config/memory_config.hpp",
-    "tt-metalium/experimental/tensor/topology/distributed_tensor_configs.hpp": "ttnn/distributed/distributed_configs.hpp",
-    "tt-metalium/experimental/tensor/topology/tensor_topology.hpp": "ttnn/distributed/tensor_topology.hpp",
+    "tt-metalium/experimental/distributed_tensor/topology/distributed_tensor_configs.hpp": "ttnn/distributed/distributed_configs.hpp",
+    "tt-metalium/experimental/distributed_tensor/topology/tensor_topology.hpp": "ttnn/distributed/tensor_topology.hpp",
 }
 
 # Files excluded from tensor forward header check - these are the forward headers themselves.
