+++ tests/tt_metal/tt_metal/api/tensor/common_tensor_test_utils.cpp
@@ -14,7 +14,7 @@
 #include <tt-metalium/experimental/tensor/mesh_tensor.hpp>
 #include <tt-metalium/experimental/tensor/tensor_apis.hpp>
 #include <tt-metalium/experimental/tensor/spec/tensor_spec.hpp>
-#include <tt-metalium/experimental/tensor/topology/tensor_topology.hpp>
+#include <tt-metalium/experimental/distributed_tensor/topology/tensor_topology.hpp>
 #include <tt_metal/impl/tensor/spec/layout/tensor_layout_impl.hpp>
 
 namespace test_utils {
