+++ tests/tt_metal/tt_metal/api/tensor/test_create_tensor_with_layout.cpp
@@ -10,7 +10,7 @@
 #include <tt-metalium/experimental/tensor/spec/layout/tensor_layout.hpp>
 #include <tt-metalium/experimental/tensor/spec/layout/page_config.hpp>
 #include <tt-metalium/experimental/tensor/tensor_types.hpp>
-#include <tt-metalium/experimental/tensor/topology/tensor_topology.hpp>
+#include <tt-metalium/experimental/distributed_tensor/topology/tensor_topology.hpp>
 
 #include "tt_metal/tt_metal/common/multi_device_fixture.hpp"
