+++ tt_metal/api/tt-metalium/experimental/distributed_tensor/topology/tensor_topology.hpp
@@ -6,7 +6,7 @@
 
 #include <tt-metalium/mesh_coord.hpp>
 
-#include <tt-metalium/experimental/tensor/topology/distributed_tensor_configs.hpp>
+#include <tt-metalium/experimental/distributed_tensor/topology/distributed_tensor_configs.hpp>
 
 namespace tt::tt_metal {
 class TensorTopology {
