+++ tt_metal/api/tt-metalium/experimental/tensor/internal/unit_tensor_apis.hpp
@@ -1,0 +1,33 @@
+// SPDX-FileCopyrightText: © 2026 Tenstorrent USA, Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include <cstddef>
+#include <optional>
+
+#include <tt-metalium/buffer.hpp>
+#include <tt-metalium/experimental/tensor/mesh_tensor.hpp>
+#include <tt-metalium/mesh_command_queue.hpp>
+
+namespace tt::tt_metal {
+
+// ======================================================================================
+//                    Unit Tensor enqueue_read/write_tensor
+// ======================================================================================
+
+void enqueue_read_tensor(
+    distributed::MeshCommandQueue& queue,
+    const MeshTensor& device_tensor,
+    std::byte* dst,
+    const std::optional<BufferRegion>& region = std::nullopt,
+    bool blocking = true);
+
+void enqueue_write_tensor(
+    distributed::MeshCommandQueue& queue,
+    const std::byte* src,
+    MeshTensor& device_tensor,
+    const std::optional<BufferRegion>& region = std::nullopt);
+
+}  // namespace tt::tt_metal
