+++ tt_metal/api/tt-metalium/experimental/tensor/mesh_tensor.hpp
@@ -8,7 +8,7 @@
 
 #include <tt-metalium/experimental/tensor/tensor_types.hpp>
 #include <tt-metalium/experimental/tensor/spec/tensor_spec.hpp>
-#include <tt-metalium/experimental/tensor/topology/tensor_topology.hpp>
+#include <tt-metalium/experimental/distributed_tensor/topology/tensor_topology.hpp>
 
 #include <tt_stl/optional_reference.hpp>
 
@@ -125,11 +125,6 @@ public:
     const TensorSpec& tensor_spec() const;
 
     /**
-     * Multi-device topology configuration - tracks how tensor is distributed across mesh devices
-     */
-    const TensorTopology& tensor_topology() const;
-
-    /**
      * Returns true if this MeshTensor was left in a moved-from state.
      *
      * A MeshTensor becomes valueless when it is the source of a move construction or move assignment.
@@ -165,11 +160,6 @@ public:
     std::size_t element_size() const;
 
     Strides strides() const;
-
-    /**
-     * Update the topology of the MeshTensor post construction.
-     */
-    void update_tensor_topology(TensorTopology tensor_topology);
 
     /**
      * Access to the implementation.
