+++ tt_metal/api/tt-metalium/experimental/tensor/tensor_apis.hpp
@@ -25,32 +25,6 @@ class MemoryConfig;
 namespace tt::tt_metal {
 
 // ======================================================================================
-//                        Transfer classification
-// ======================================================================================
-
-// Returns true if a H2D between the HostTensor and target MeshDevice is uniform.
-// A transfer is uniform if the host shards cover the entire shape of the MeshDevice.
-//
-// Example of uniform transfer:
-// HostTensor with a DistributedHostBuffer of shards [0, 0], [0, 1], [1, 0], [1, 1] (shape 2x2).
-// MeshDevice of shape 2x2.
-// Here the shards map exactly to the shape of the MeshDevice.
-//
-// Example of non-uniform transfers:
-//
-// 1: one to many replicas
-// HostTensor with a single shard at [0,0].
-// MeshDevice of shape 2x2.
-// This is a replica-based non-uniform transfer.
-//
-// 2: partial coverage:
-// HostTensor with a DistributedHostBuffer of shards [0, 0], and [1, 0]
-// MeshDevice of shape 2x2.
-// This is a partial coverage non-uniform transfer.
-// Only opposite sides of the MeshDevice will receive new data.
-bool is_uniform_write(const HostTensor& host_tensor, const distributed::MeshDevice& device);
-
-// ======================================================================================
 //                   Uniform enqueue_read/write_tensor
 // ======================================================================================
 
@@ -67,56 +41,6 @@ MeshTensor enqueue_write_tensor(
     ttsl::optional_reference<const MemoryConfig> memory_config = std::nullopt);
 
 void enqueue_write_tensor(distributed::MeshCommandQueue& cq, const HostTensor& host_tensor, MeshTensor& device_tensor);
-
-// ======================================================================================
-//                    Unit Tensor enqueue_read/write_tensor
-// ======================================================================================
-
-void enqueue_read_tensor(
-    distributed::MeshCommandQueue& queue,
-    const MeshTensor& device_tensor,
-    std::byte* dst,
-    const std::optional<BufferRegion>& region = std::nullopt,
-    bool blocking = true);
-
-void enqueue_write_tensor(
-    distributed::MeshCommandQueue& queue,
-    const std::byte* src,
-    MeshTensor& device_tensor,
-    const std::optional<BufferRegion>& region = std::nullopt);
-
-// ======================================================================================
-//                Non-uniform enqueue_read/write_tensor
-// ======================================================================================
-
-// Data movement for tensors whose shards don't cover the entire MeshDevice.
-// The host-side DistributedHostBuffer only populates a subset of MeshCoordinates,
-// so the resulting DeviceStorage must track which coordinates were actually written.
-namespace non_uniform_data_movement {
-
-HostTensor enqueue_read_tensor(
-    distributed::MeshCommandQueue& cq,
-    const MeshTensor& device_tensor,
-    std::span<const distributed::MeshCoordinate> coords,
-    bool blocking = true);
-
-void enqueue_read_tensor(
-    distributed::MeshCommandQueue& cq,
-    const MeshTensor& device_tensor,
-    HostTensor& host_tensor,
-    std::span<const distributed::MeshCoordinate> coords,
-    bool blocking = true);
-
-std::pair<MeshTensor, std::vector<distributed::MeshCoordinate>> enqueue_write_tensor(
-    distributed::MeshCommandQueue& cq,
-    const HostTensor& host_tensor,
-    distributed::MeshDevice& mesh_device,
-    ttsl::optional_reference<const MemoryConfig> memory_config = std::nullopt);
-
-std::vector<distributed::MeshCoordinate> enqueue_write_tensor(
-    distributed::MeshCommandQueue& cq, const HostTensor& host_tensor, MeshTensor& device_tensor);
-
-}  // namespace non_uniform_data_movement
 
 // ======================================================================================
 //                                  .to_layout()
