+++ tt_metal/impl/experimental/core_subset_write/tensor.cpp
@@ -5,6 +5,7 @@
 #include <tt-metalium/experimental/core_subset_write/mesh_command_queue.hpp>
 #include <tt-metalium/experimental/core_subset_write/tensor.hpp>
 #include <tt-metalium/experimental/tensor/tensor_apis.hpp>
+#include <tt-metalium/experimental/distributed_tensor/distributed_tensor_apis.hpp>
 
 #include "tt_metal/impl/tensor/mesh_tensor_impl.hpp"
