+++ tt_metal/impl/tensor/tensor_apis.cpp
@@ -15,6 +15,7 @@
 #include "mesh_tensor_impl.hpp"
 
 #include <tt-metalium/bfloat16.hpp>
+#include <tt-metalium/experimental/distributed_tensor/distributed_tensor_apis.hpp>
 #include <tt-metalium/experimental/tensor/tensor_apis.hpp>
 #include <tt-metalium/experimental/tensor/impl/tensor_impl.hpp>
 #include <tt-metalium/experimental/tensor/tensor_types.hpp>
@@ -31,23 +32,6 @@
 #include <tt_stl/span.hpp>
 
 namespace tt::tt_metal {
-
-// ======================================================================================
-//                            Transfer classification
-// ======================================================================================
-
-bool is_uniform_write(const HostTensor& host_tensor, const distributed::MeshDevice& device) {
-    const auto& device_mesh_shape = device.shape();
-    const auto& host_buffer = host_tensor.buffer();
-
-    if (host_buffer.shape() != device_mesh_shape) {
-        return false;
-    }
-
-    auto all_coords = distributed::MeshCoordinateRange(device_mesh_shape);
-    return std::ranges::all_of(
-        all_coords, [&](const auto& coord) { return host_buffer.shard_coords().contains(coord); });
-}
 
 namespace {
 namespace CMAKE_UNIQUE_NAMESPACE {
@@ -88,7 +72,7 @@ HostTensor enqueue_read_tensor(distributed::MeshCommandQueue& cq, const MeshTens
     cq.enqueue_read(mesh_buffer, distributed_host_buffer, /*shards=*/std::nullopt, blocking);
 
     return HostTensor::from_buffer(
-        std::move(distributed_host_buffer), device_tensor.tensor_spec(), device_tensor.tensor_topology());
+        std::move(distributed_host_buffer), device_tensor.tensor_spec(), tensor_topology(device_tensor));
 }
 
 MeshTensor enqueue_write_tensor(
@@ -116,7 +100,7 @@ MeshTensor enqueue_write_tensor(
                                   ? &tensor_spec_overriden_memory_config.value()
                                   : &host_tensor.tensor_spec();
 
-    auto result = MeshTensor::allocate_on_device(mesh_device, *tensor_spec, host_tensor.tensor_topology());
+    auto result = MeshTensor::allocate_on_device(mesh_device, *tensor_spec, tensor_topology(host_tensor));
     enqueue_write_tensor(cq, host_tensor, result);
     return result;
 }
@@ -158,7 +142,7 @@ void enqueue_read_tensor(
     }
 
     cq.enqueue_read(mesh_buffer, dst_distributed_host_buffer, /*shards=*/std::nullopt, blocking);
-    host_tensor.update_tensor_topology(device_tensor.tensor_topology());
+    update_tensor_topology(host_tensor, tensor_topology(device_tensor));
 }
 
 void enqueue_write_tensor(distributed::MeshCommandQueue& cq, const HostTensor& host_tensor, MeshTensor& device_tensor) {
@@ -239,314 +223,8 @@ void enqueue_write_tensor(distributed::MeshCommandQueue& cq, const HostTensor& h
                 host_tensor.tensor_spec().tensor_layout().get_page_config(),
                 device_tensor.memory_config(),
                 host_tensor.tensor_spec().tensor_layout().get_alignment())),
-        host_tensor.tensor_topology());
+        tensor_topology(host_tensor));
 }
-
-// ======================================================================================
-//                    Unit Tensor enqueue_read/write_tensor
-// ======================================================================================
-
-void enqueue_read_tensor(
-    distributed::MeshCommandQueue& queue,
-    const MeshTensor& device_tensor,
-    std::byte* dst,
-    const std::optional<BufferRegion>& region,
-    bool blocking) {
-    TT_FATAL(queue.device()->num_devices() == 1, "enqueue_read_tensor only supports single device mesh");
-    std::vector<distributed::ShardDataTransfer> shard_data_transfers = {
-        distributed::ShardDataTransfer{*distributed::MeshCoordinateRange(queue.device()->shape()).begin()}
-            .host_data(dst)
-            .region(region)};
-    queue.enqueue_read_shards(shard_data_transfers, device_tensor.impl().raw_mesh_buffer(), blocking);
-}
-
-void enqueue_write_tensor(
-    distributed::MeshCommandQueue& queue,
-    const std::byte* src,
-    MeshTensor& device_tensor,
-    const std::optional<BufferRegion>& region) {
-    TT_FATAL(queue.device()->num_devices() == 1, "enqueue_write_tensor only supports single device mesh");
-    std::vector<distributed::ShardDataTransfer> shard_data_transfers = {
-        distributed::ShardDataTransfer{*distributed::MeshCoordinateRange(queue.device()->shape()).begin()}
-            .host_data(const_cast<std::byte*>(src))
-            .region(region)};
-    queue.enqueue_write_shards(device_tensor.impl().raw_mesh_buffer(), shard_data_transfers, false);
-}
-
-// ======================================================================================
-//              Non-uniform enqueue_read/write_tensor
-// ======================================================================================
-
-namespace non_uniform_data_movement {
-
-HostTensor enqueue_read_tensor(
-    distributed::MeshCommandQueue& cq,
-    const MeshTensor& device_tensor,
-    std::span<const distributed::MeshCoordinate> coords,
-    bool blocking) {
-    auto distributed_host_buffer = DistributedHostBuffer::create(device_tensor.device().get_view());
-    distributed_host_buffer.emplace_shards(
-        {coords.begin(), coords.end()},
-        [&](const distributed::MeshCoordinate&) {
-            return tensor_impl::allocate_host_buffer(device_tensor.tensor_spec());
-        },
-        DistributedHostBuffer::ProcessShardExecutionPolicy::PARALLEL);
-
-    auto result = HostTensor::from_buffer(
-        std::move(distributed_host_buffer), device_tensor.tensor_spec(), device_tensor.tensor_topology());
-    enqueue_read_tensor(cq, device_tensor, result, coords, blocking);
-    return result;
-}
-
-void enqueue_read_tensor(
-    distributed::MeshCommandQueue& cq,
-    const MeshTensor& device_tensor,
-    HostTensor& host_tensor,
-    std::span<const distributed::MeshCoordinate> coords,
-    bool blocking) {
-    TT_FATAL(host_tensor.logical_shape() == device_tensor.logical_shape(), "Host tensor has different shape");
-    TT_FATAL(host_tensor.dtype() == device_tensor.dtype(), "Host tensor has different dtype");
-    TT_FATAL(
-        host_tensor.tensor_spec().page_config() == device_tensor.tensor_spec().page_config(),
-        "Host tensor has different page config");
-
-    const auto& distributed_host_buffer = host_tensor.buffer();
-
-    std::vector<std::pair<distributed::MeshCoordinate, std::optional<HostBuffer>>> shards;
-    shards.reserve(coords.size());
-    for (const auto& device_coord : coords) {
-        shards.push_back({device_coord, distributed_host_buffer.get_shard(device_coord)});
-    }
-
-    DistributedHostBuffer dst_distributed_host_buffer =
-        DistributedHostBuffer::create(device_tensor.device().get_view());
-    const size_t expected_size_bytes = device_tensor.tensor_spec().compute_packed_buffer_size_bytes();
-    for (const auto& [device_coord, host_buffer] : shards) {
-        dst_distributed_host_buffer.emplace_shard(device_coord, [&]() {
-            TT_FATAL(host_buffer.has_value(), "Host shard for device shard {} is not populated.", device_coord);
-            TT_FATAL(
-                host_buffer->view_bytes().size() == expected_size_bytes,
-                "Host shard for device shard {} has invalid size: {} != {}",
-                device_coord,
-                host_buffer->view_bytes().size(),
-                expected_size_bytes);
-            return *host_buffer;
-        });
-    }
-
-    std::unordered_set<distributed::MeshCoordinate> shard_set(coords.begin(), coords.end());
-    cq.enqueue_read(device_tensor.impl().raw_mesh_buffer(), dst_distributed_host_buffer, shard_set, blocking);
-
-    host_tensor = HostTensor::from_buffer(
-        std::move(dst_distributed_host_buffer), device_tensor.tensor_spec(), device_tensor.tensor_topology());
-}
-
-std::pair<MeshTensor, std::vector<distributed::MeshCoordinate>> enqueue_write_tensor(
-    distributed::MeshCommandQueue& cq,
-    const HostTensor& host_tensor,
-    distributed::MeshDevice& mesh_device,
-    ttsl::optional_reference<const MemoryConfig> memory_config) {
-    std::optional<TensorSpec> tensor_spec_overriden_memory_config;
-    if (memory_config) {
-        const auto& old_spec = host_tensor.tensor_spec();
-        tensor_spec_overriden_memory_config = TensorSpec(
-            old_spec.logical_shape(),
-            TensorLayout(
-                old_spec.tensor_layout().get_data_type(),
-                old_spec.tensor_layout().get_page_config(),
-                *memory_config,
-                old_spec.tensor_layout().get_alignment()));
-    }
-
-    const auto* tensor_spec = tensor_spec_overriden_memory_config.has_value()
-                                  ? &tensor_spec_overriden_memory_config.value()
-                                  : &host_tensor.tensor_spec();
-
-    auto result = MeshTensor::allocate_on_device(mesh_device, *tensor_spec, host_tensor.tensor_topology());
-    auto coords = non_uniform_data_movement::enqueue_write_tensor(cq, host_tensor, result);
-    return {std::move(result), std::move(coords)};
-}
-
-namespace {
-namespace CMAKE_UNIQUE_NAMESPACE {
-
-void h2d_as_replicate_tensor_on_1x1_mesh(
-    const HostTensor& host_tensor, MeshTensor& device_tensor, distributed::MeshCommandQueue& command_queue) {
-    const auto host_buffer = host_tensor.buffer().get_shard(distributed::MeshCoordinate(0, 0));
-    auto data_to_write = host_buffer->view_bytes();
-    const auto expected_packed_buffer_size_bytes = device_tensor.tensor_spec().compute_packed_buffer_size_bytes();
-    const auto input_size_bytes = data_to_write.size();
-    TT_FATAL(
-        input_size_bytes == expected_packed_buffer_size_bytes,
-        "Host data with total size {}B does not match expected size {}B of device buffer!",
-        input_size_bytes,
-        expected_packed_buffer_size_bytes);
-
-    auto mesh_buffer = device_tensor.impl().raw_mesh_buffer();
-    auto* mesh_device = mesh_buffer->device();
-
-    const bool use_pinned =
-        ::tt::tt_metal::CMAKE_UNIQUE_NAMESPACE::should_use_pinned_write_path(*mesh_device, data_to_write.size());
-
-    if (use_pinned) {
-        // Replication fans a single 1x1 host shard out to the whole mesh, but only the chips owned
-        // by this host can be pinned/written; restrict the pin and the transfers to those so the
-        // remote coordinates are a complete no-op here.
-        const auto& view = mesh_device->get_view();
-        std::vector<distributed::MeshCoordinate> local_coords;
-        distributed::MeshCoordinateRangeSet local_range;
-        for (const auto& coord : distributed::MeshCoordinateRange(mesh_device->shape())) {
-            if (view.impl().is_local(coord)) {
-                local_coords.push_back(coord);
-                local_range.merge(distributed::MeshCoordinateRange(coord, coord));
-            }
-        }
-
-        HostBuffer pinned_buffer(*host_buffer);
-        auto pinned_memory = local_coords.empty() ? nullptr
-                                                  : experimental::PinnedMemoryCache::instance().try_pin(
-                                                        *mesh_device, local_range, pinned_buffer, /*map_to_noc=*/true);
-
-        if (pinned_memory) {
-            std::vector<distributed::ShardDataTransfer> transfers;
-            transfers.reserve(local_coords.size());
-            for (const auto& coord : local_coords) {
-                auto xfer = distributed::ShardDataTransfer{coord}
-                                .host_data(const_cast<void*>(static_cast<const void*>(data_to_write.data())))
-                                .region(BufferRegion(0, data_to_write.size()));
-                experimental::ShardDataTransferSetPinnedMemory(xfer, pinned_memory);
-                transfers.push_back(std::move(xfer));
-            }
-            command_queue.enqueue_write_shards(mesh_buffer, transfers, /*blocking=*/true);
-        } else {
-            command_queue.enqueue_write_mesh_buffer(mesh_buffer, data_to_write.data(), /*blocking=*/false);
-        }
-    } else {
-        command_queue.enqueue_write_mesh_buffer(mesh_buffer, data_to_write.data(), /*blocking=*/false);
-    }
-
-    const auto& mesh_device_shape = mesh_buffer->device()->shape();
-    auto topology = TensorTopology::create_fully_replicated_tensor_topology(mesh_device_shape);
-    const auto& old_spec = host_tensor.tensor_spec();
-    device_tensor = MeshTensor::from_buffer(
-        std::move(*mesh_buffer),
-        TensorSpec(
-            old_spec.logical_shape(),
-            TensorLayout(
-                old_spec.tensor_layout().get_data_type(),
-                old_spec.tensor_layout().get_page_config(),
-                device_tensor.memory_config(),
-                old_spec.tensor_layout().get_alignment())),
-        topology);
-}
-
-}  // namespace CMAKE_UNIQUE_NAMESPACE
-}  // namespace
-
-std::vector<distributed::MeshCoordinate> enqueue_write_tensor(
-    distributed::MeshCommandQueue& cq, const HostTensor& host_tensor, MeshTensor& device_tensor) {
-    TT_FATAL(host_tensor.logical_shape() == device_tensor.logical_shape(), "Host tensor has different shape");
-    TT_FATAL(host_tensor.dtype() == device_tensor.dtype(), "Host tensor has different dtype");
-    TT_FATAL(
-        host_tensor.tensor_spec().page_config() == device_tensor.tensor_spec().page_config(),
-        "Host tensor has different page config");
-
-    const auto& host_storage_shape = host_tensor.buffer().shape();
-    const auto& dst_device_shape = device_tensor.device().shape();
-
-    // Special case of replicating tensors on 1x1 mesh across the entire mesh device.
-    if (host_storage_shape.mesh_size() < dst_device_shape.mesh_size() &&
-        host_storage_shape == distributed::MeshShape(1, 1)) {
-        CMAKE_UNIQUE_NAMESPACE::h2d_as_replicate_tensor_on_1x1_mesh(host_tensor, device_tensor, cq);
-
-        // All coordinates of the MeshDevice
-        distributed::MeshCoordinateRange range(device_tensor.device().shape());
-        return {range.begin(), range.end()};
-    }
-
-    auto mesh_buffer = device_tensor.impl().raw_mesh_buffer();
-    const auto& distributed_host_buffer = host_tensor.buffer();
-
-    size_t total_size = 0;
-    for (const auto& coord : distributed_host_buffer.shard_coords()) {
-        auto buf = distributed_host_buffer.get_shard(coord);
-        if (buf) {
-            total_size += buf->view_bytes().size();
-        }
-    }
-
-    const bool use_pinned =
-        ::tt::tt_metal::CMAKE_UNIQUE_NAMESPACE::should_use_pinned_write_path(*cq.device(), total_size);
-
-    if (use_pinned) {
-        auto* mesh_device = mesh_buffer->device();
-        const auto& view = mesh_device->get_view();
-        std::vector<distributed::ShardDataTransfer> transfers;
-        transfers.reserve(distributed_host_buffer.shard_coords().size());
-        bool any_pinned = false;
-
-        for (const auto& coord : distributed_host_buffer.shard_coords()) {
-            // get_shard yields a buffer only for shards owned by this host, so remote chips are
-            // never pinned or added to the transfer list -- the transfer is a no-op for them here.
-            auto buf = distributed_host_buffer.get_shard(coord);
-            if (buf) {
-                // The host buffer's distribution must agree with the device's: host memory can only
-                // be pinned to MMIO devices local to this process, so a populated shard for a coord
-                // the device owns on another host must never reach try_pin (which would fault while
-                // resolving the remote device).
-                TT_FATAL(
-                    view.impl().is_local(coord),
-                    "Host buffer holds a shard for device coordinate {}, but that device is not local "
-                    "to this host; host memory can only be pinned to MMIO devices owned by this process.",
-                    coord);
-                auto coord_range = distributed::MeshCoordinateRangeSet(distributed::MeshCoordinateRange(coord, coord));
-                HostBuffer pinned_buf(*buf);
-                auto pinned_memory = experimental::PinnedMemoryCache::instance().try_pin(
-                    *mesh_device, coord_range, pinned_buf, /*map_to_noc=*/true);
-
-                auto xfer = distributed::ShardDataTransfer{distributed::MeshCoordinate(coord)}
-                                .host_data(buf->view_bytes().data())
-                                .region(BufferRegion(0, buf->view_bytes().size()));
-                if (pinned_memory) {
-                    experimental::ShardDataTransferSetPinnedMemory(xfer, std::move(pinned_memory));
-                    any_pinned = true;
-                }
-                transfers.push_back(std::move(xfer));
-            }
-        }
-        if (any_pinned) {
-            cq.enqueue_write_shards(mesh_buffer, transfers, /*blocking=*/true);
-        } else {
-            cq.enqueue_write(mesh_buffer, distributed_host_buffer, /*blocking=*/false);
-        }
-    } else {
-        cq.enqueue_write(mesh_buffer, distributed_host_buffer, /*blocking=*/false);
-    }
-
-    // DistributedHostBuffer may not cover the entire MeshDevice, must preserve coords here.
-    // Coordinates here represents the shards that are local to this instance, there maybe other shards that are on
-    // another host.
-    std::vector<distributed::MeshCoordinate> coords;
-    const auto& shard_coords = distributed_host_buffer.shard_coords();
-    coords.reserve(shard_coords.size());
-    std::copy(shard_coords.begin(), shard_coords.end(), std::back_inserter(coords));
-
-    const auto& old_spec = host_tensor.tensor_spec();
-    device_tensor = MeshTensor::from_buffer(
-        std::move(*mesh_buffer),
-        TensorSpec(
-            old_spec.logical_shape(),
-            TensorLayout(
-                old_spec.tensor_layout().get_data_type(),
-                old_spec.tensor_layout().get_page_config(),
-                device_tensor.memory_config(),
-                old_spec.tensor_layout().get_alignment())),
-        host_tensor.tensor_topology());
-
-    return coords;
-}
-
-}  // namespace non_uniform_data_movement
 
 // ======================================================================================
 //                                  .to_layout()
@@ -599,7 +277,8 @@ HostTensor to_row_major_layout_impl(const HostTensor& tensor) {
         },
         DistributedHostBuffer::ProcessShardExecutionPolicy::PARALLEL);
 
-    return HostTensor::from_buffer(std::move(transformed_buffer), new_tensor_spec, tensor.tensor_topology());
+    return HostTensor::from_buffer(std::move(transformed_buffer), new_tensor_spec, tensor_topology(tensor));
+
 }
 
 template <typename T>
@@ -642,7 +321,7 @@ HostTensor to_tile_layout_impl(const HostTensor& tensor, Tile tile) {
 
         assert_host_shards_match_packed_size(transformed_buffer, output_spec, "to_tile_layout");
 
-        return HostTensor::from_buffer(std::move(transformed_buffer), output_spec, tensor.tensor_topology());
+        return HostTensor::from_buffer(std::move(transformed_buffer), output_spec, tensor_topology(tensor));
     }
 }
 
@@ -727,7 +406,7 @@ HostTensor pad_bfloat8_b(
                 MemoryConfig{},
                 tensor.logical_shape(),
                 tensor.padded_shape())),
-        tensor.tensor_topology());
+        tensor_topology(tensor));
     auto float_tensor = pad(intermediate, output_padded_shape, input_tensor_start, pad_value);
 
     // Convert back to BFLOAT8_B
@@ -743,7 +422,7 @@ HostTensor pad_bfloat8_b(
             MemoryConfig{},
             float_tensor.logical_shape(),
             float_tensor.padded_shape()));
-    return HostTensor::from_buffer(std::move(output_uint32_buffer), output_spec, tensor.tensor_topology());
+    return HostTensor::from_buffer(std::move(output_uint32_buffer), output_spec, tensor_topology(tensor));
 }
 
 HostTensor unpad_bfloat8_b(
@@ -769,7 +448,7 @@ HostTensor unpad_bfloat8_b(
                 MemoryConfig{},
                 tensor.logical_shape(),
                 tensor.padded_shape())),
-        tensor.tensor_topology());
+        tensor_topology(tensor));
     auto float_tensor = unpad(intermediate, output_tensor_start, output_tensor_end);
 
     // Convert back to BFLOAT8_B
@@ -787,7 +466,7 @@ HostTensor unpad_bfloat8_b(
                 MemoryConfig{},
                 float_tensor.logical_shape(),
                 float_tensor.padded_shape())),
-        tensor.tensor_topology());
+        tensor_topology(tensor));
 }
 
 HostTensor pad_bfloat4_b(
@@ -813,7 +492,7 @@ HostTensor pad_bfloat4_b(
                 MemoryConfig{},
                 tensor.logical_shape(),
                 tensor.logical_shape())),
-        tensor.tensor_topology());
+        tensor_topology(tensor));
     auto float_tensor = pad(intermediate, output_padded_shape, input_tensor_start, pad_value);
 
     // Convert back to BFLOAT4_B
@@ -829,7 +508,7 @@ HostTensor pad_bfloat4_b(
             MemoryConfig{},
             float_tensor.logical_shape(),
             float_tensor.padded_shape()));
-    return HostTensor::from_buffer(std::move(output_uint32_buffer), output_spec, tensor.tensor_topology());
+    return HostTensor::from_buffer(std::move(output_uint32_buffer), output_spec, tensor_topology(tensor));
 }
 
 HostTensor unpad_bfloat4_b(
@@ -854,7 +533,7 @@ HostTensor unpad_bfloat4_b(
                 MemoryConfig{},
                 tensor.logical_shape(),
                 tensor.padded_shape())),
-        tensor.tensor_topology());
+        tensor_topology(tensor));
     auto float_tensor = unpad(intermediate, output_tensor_start, output_tensor_end);
 
     // Convert back to BFLOAT4_B
@@ -872,7 +551,7 @@ HostTensor unpad_bfloat4_b(
                 MemoryConfig{},
                 float_tensor.logical_shape(),
                 float_tensor.padded_shape())),
-        tensor.tensor_topology());
+        tensor_topology(tensor));
 }
 
 template <typename T>
@@ -971,7 +650,7 @@ HostTensor pad_impl(
                 MemoryConfig{},
                 tensor.logical_shape(),
                 output_padded_shape)),
-        tensor.tensor_topology());
+        tensor_topology(tensor));
 }
 
 template <>
@@ -1055,7 +734,7 @@ HostTensor unpad_impl(
                 tensor.dtype(),
                 tt::tt_metal::PageConfig(tensor.layout(), tensor.tensor_spec().tile()),
                 tt::tt_metal::MemoryConfig{})),
-        tensor.tensor_topology());
+        tensor_topology(tensor));
 }
 
 template <>
@@ -1351,7 +1030,7 @@ HostTensor to_dtype(const HostTensor& input_tensor, DataType dtype) {
         }
     }
 
-    return HostTensor::from_buffer(std::move(result_buffer), output_spec, input_tensor.tensor_topology());
+    return HostTensor::from_buffer(std::move(result_buffer), output_spec, tensor_topology(input_tensor));
 }
 
 // ======================================================================================
@@ -1483,7 +1162,7 @@ HostTensor to_tensor_spec(const HostTensor& tensor, const TensorSpec& dest_spec,
 
     CMAKE_UNIQUE_NAMESPACE::assert_packed_shard_sizes(transformed_buffer, working_spec);
 
-    HostTensor result = HostTensor::from_buffer(std::move(transformed_buffer), working_spec, tensor.tensor_topology());
+    HostTensor result = HostTensor::from_buffer(std::move(transformed_buffer), working_spec, tensor_topology(tensor));
 
     if (result.dtype() != dest_spec.data_type()) {
         result = to_dtype(result, dest_spec.data_type());
@@ -1493,7 +1172,7 @@ HostTensor to_tensor_spec(const HostTensor& tensor, const TensorSpec& dest_spec,
         CMAKE_UNIQUE_NAMESPACE::exact_spec_match(result.tensor_spec(), dest_spec),
         "to_tensor_spec: result does not satisfy exact-spec predicate against dest_spec");
     CMAKE_UNIQUE_NAMESPACE::assert_packed_shard_sizes(result.buffer(), dest_spec);
-    TT_FATAL(result.tensor_topology() == tensor.tensor_topology(), "to_tensor_spec: topology must be preserved");
+    TT_FATAL(tensor_topology(result) == tensor_topology(tensor), "to_tensor_spec: topology must be preserved");
 
     return result;
 }
