+++ tt_metal/impl/tensor/unit_tensor_apis.cpp
@@ -1,0 +1,38 @@
+// SPDX-FileCopyrightText: © 2026 Tenstorrent USA, Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#include <tt-metalium/experimental/tensor/internal/unit_tensor_apis.hpp>
+
+#include "mesh_tensor_impl.hpp"
+
+namespace tt::tt_metal {
+
+void enqueue_read_tensor(
+    distributed::MeshCommandQueue& queue,
+    const MeshTensor& device_tensor,
+    std::byte* dst,
+    const std::optional<BufferRegion>& region,
+    bool blocking) {
+    TT_FATAL(queue.device()->num_devices() == 1, "enqueue_read_tensor only supports single device mesh");
+    std::vector<distributed::ShardDataTransfer> shard_data_transfers = {
+        distributed::ShardDataTransfer{*distributed::MeshCoordinateRange(queue.device()->shape()).begin()}
+            .host_data(dst)
+            .region(region)};
+    queue.enqueue_read_shards(shard_data_transfers, device_tensor.impl().raw_mesh_buffer(), blocking);
+}
+
+void enqueue_write_tensor(
+    distributed::MeshCommandQueue& queue,
+    const std::byte* src,
+    MeshTensor& device_tensor,
+    const std::optional<BufferRegion>& region) {
+    TT_FATAL(queue.device()->num_devices() == 1, "enqueue_write_tensor only supports single device mesh");
+    std::vector<distributed::ShardDataTransfer> shard_data_transfers = {
+        distributed::ShardDataTransfer{*distributed::MeshCoordinateRange(queue.device()->shape()).begin()}
+            .host_data(const_cast<std::byte*>(src))
+            .region(region)};
+    queue.enqueue_write_shards(device_tensor.impl().raw_mesh_buffer(), shard_data_transfers, false);
+}
+
+}  // namespace tt::tt_metal
