+++ ttnn/api/ttnn/distributed/distributed_configs.hpp
@@ -4,4 +4,4 @@
 
 #pragma once
 
-#include <tt-metalium/experimental/tensor/topology/distributed_tensor_configs.hpp>
+#include <tt-metalium/experimental/distributed_tensor/topology/distributed_tensor_configs.hpp>
