+++ ttnn/core/tensor/tensor_ops.cpp
@@ -22,6 +22,8 @@
 #include "ttnn/graph/graph_serialization.hpp"
 
 #include <tt-metalium/experimental/tensor/tensor_apis.hpp>
+#include <tt-metalium/experimental/distributed_tensor/distributed_tensor_apis.hpp>
+#include <tt-metalium/experimental/tensor/internal/unit_tensor_apis.hpp>
 
 using tt::tt_metal::BufferRegion;
 using tt::tt_metal::DataType;
