+++ ttnn/cpp/ttnn/operations/experimental/core_subset_write/copy_to_device_filtered.cpp
@@ -10,6 +10,7 @@
 
 #include <tt-metalium/experimental/core_subset_write/tensor.hpp>
 #include <tt-metalium/experimental/tensor/tensor_apis.hpp>
+#include <tt-metalium/experimental/distributed_tensor/distributed_tensor_apis.hpp>
 #include <tt-metalium/queue_id.hpp>
 
 #include <tracy/Tracy.hpp>
